Big refactor for Handlers.
