+++ src/components/core/downloadHandler.ts
@@ -1,35 +1,109 @@
 import { JsonRpcProvider } from 'ethers'
-import {
-  DownloadTask,
-  PROTOCOL_COMMANDS,
-  DownloadURLCommand
-} from '../../utils/constants.js'
-import { Handler, FindDdoHandler } from './handler.js'
+import { DownloadTask } from '../../utils/constants.js'
+import { Handler } from './handler.js'
 import { checkNonce, NonceResponse } from './utils/nonceHandler.js'
-import { OceanNodeConfig, P2PCommandResponse } from '../../@types/OceanNode.js'
+import { P2PCommandResponse } from '../../@types/OceanNode.js'
 import { P2P_CONSOLE_LOGGER } from '../P2P/index.js'
 import { validateOrderTransaction } from './validateTransaction.js'
 import { AssetUtils } from '../../utils/asset.js'
 import { Service } from '../../@types/DDO/Service.js'
 import { checkFee } from './utils/feesHandler.js'
 import { decrypt } from '../../utils/crypt.js'
-import { Database } from '../database/index.js'
+import { FindDdoHandler } from './ddoHandler.js'
 import crypto from 'crypto'
 import * as ethCrypto from 'eth-crypto'
 import { GENERIC_EMOJIS, LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
 import { Storage } from '../storage/index.js'
 export const FILE_ENCRYPTION_ALGORITHM = 'aes-256-cbc'
 
-export class DownloadHandler extends Handler {
-  public constructor(task: any, config: OceanNodeConfig, db: Database) {
-    super(task, config, db)
-    if (!this.isDownloadCommand(task) && !this.isDownloadUrlCommand(task)) {
-      throw new Error(
-        `Task has nor DownloadCommand type, nor DownloadUrlCommand type. It has ${typeof task}`
+export async function handleDownloadUrlCommand(
+  task: any,
+  config: any
+): Promise<P2PCommandResponse> {
+  const encryptFile = !!task.aes_encrypted_key
+  P2P_CONSOLE_LOGGER.logMessage(
+    'DownloadUrlCommand requires file encryption? ' + encryptFile,
+    true
+  )
+
+  try {
+    // Determine the type of storage and get a readable stream
+    const storage = Storage.getStorageClass(task.fileObject)
+    const inputStream = await storage.getReadableStream()
+    const headers: any = {}
+    for (const [key, value] of Object.entries(inputStream.headers)) {
+      headers[key] = value
+    }
+    // need to check if content length is already in headers, but we don't know the case
+    const objTemp = JSON.parse(JSON.stringify(headers).toLowerCase())
+    if (!('Content-Length'.toLowerCase() in objTemp))
+      headers['Transfer-Encoding'] = 'chunked'
+    if (!('Content-Disposition'.toLowerCase() in objTemp))
+      headers['Content-Disposition'] = 'attachment;filename=unknownfile' // TO DO: use did+serviceId+fileIndex
+    if (encryptFile) {
+      // we parse the string into the object again
+      const encryptedObject = ethCrypto.cipher.parse(task.aes_encrypted_key)
+      // get the key from configuration
+      const nodePrivateKey = Buffer.from(config.keys.privateKey).toString('hex')
+      const decrypted = await ethCrypto.decryptWithPrivateKey(
+        nodePrivateKey,
+        encryptedObject
       )
+      const decryptedPayload = JSON.parse(decrypted)
+      // check signature
+      // const senderAddress = ethCrypto.recover(
+      //  decryptedPayload.signature,
+      //  ethCrypto.hash.keccak256(decryptedPayload.message)
+      // )
+      // Optional, we can also validate the original address of the sender (the client that created the message)
+      // this could be part of the /directCommand payload for instance
+      // console.log(
+      //  'Got message from ' + senderAddress + ' secrets: ' + decryptedPayload.message
+      // )
+      const secrets = JSON.parse(decryptedPayload.message)
+
+      const cipher = crypto
+        .createCipheriv(
+          FILE_ENCRYPTION_ALGORITHM,
+          Buffer.from(secrets.key, 'hex'),
+          Buffer.from(secrets.iv, 'hex')
+        )
+        .setAutoPadding(true)
+
+      headers['Content-Encoding'] = 'aesgcm'
+
+      return {
+        stream: inputStream.stream.pipe(cipher),
+        status: {
+          httpStatus: inputStream.httpStatus,
+          headers
+        }
+      }
+    } else {
+      // Download request is not using encryption!
+      return {
+        stream: inputStream.stream,
+        status: {
+          httpStatus: inputStream.httpStatus,
+          headers
+        }
+      }
+    }
+  } catch (err) {
+    P2P_CONSOLE_LOGGER.logMessageWithEmoji(
+      'Failure executing downloadURL task: ' + err.message,
+      true,
+      GENERIC_EMOJIS.EMOJI_CROSS_MARK,
+      LOG_LEVELS_STR.LEVEL_ERROR
+    )
+    return {
+      stream: null,
+      status: { httpStatus: 501, error: 'Unknown error: ' + err.message }
     }
   }
+}
 
+export class DownloadHandler extends Handler {
   isDownloadCommand(obj: any): obj is DownloadTask {
     return (
       typeof obj === 'object' &&
@@ -44,102 +118,14 @@ export class DownloadHandler extends Handler {
     )
   }
 
-  isDownloadUrlCommand(obj: any): obj is DownloadURLCommand {
-    return (
-      typeof obj === 'object' && obj !== null && 'command' in obj && 'fileObject' in obj
-    )
-  }
-
   // No encryption here yet
-  // Put it in DownloadHandler to avoid ReferenceError: Cannot access 'Handler' before initialization
-  async handleDownloadUrlCommand(): Promise<P2PCommandResponse> {
-    const task = this.getTask()
-    const config = this.getConfig()
-    const encryptFile = !!task.aes_encrypted_key
-    P2P_CONSOLE_LOGGER.logMessage(
-      'DownloadUrlCommand requires file encryption? ' + encryptFile,
-      true
-    )
 
-    try {
-      // Determine the type of storage and get a readable stream
-      const storage = Storage.getStorageClass(task.fileObject)
-      const inputStream = await storage.getReadableStream()
-      const headers: any = {}
-      for (const [key, value] of Object.entries(inputStream.headers)) {
-        headers[key] = value
-      }
-      // need to check if content length is already in headers, but we don't know the case
-      const objTemp = JSON.parse(JSON.stringify(headers).toLowerCase())
-      if (!('Content-Length'.toLowerCase() in objTemp))
-        headers['Transfer-Encoding'] = 'chunked'
-      if (!('Content-Disposition'.toLowerCase() in objTemp))
-        headers['Content-Disposition'] = 'attachment;filename=unknownfile' // TO DO: use did+serviceId+fileIndex
-      if (encryptFile) {
-        // we parse the string into the object again
-        const encryptedObject = ethCrypto.cipher.parse(task.aes_encrypted_key)
-        // get the key from configuration
-        const nodePrivateKey = Buffer.from(config.keys.privateKey).toString('hex')
-        const decrypted = await ethCrypto.decryptWithPrivateKey(
-          nodePrivateKey,
-          encryptedObject
-        )
-        const decryptedPayload = JSON.parse(decrypted)
-        // check signature
-        // const senderAddress = ethCrypto.recover(
-        //  decryptedPayload.signature,
-        //  ethCrypto.hash.keccak256(decryptedPayload.message)
-        // )
-        // Optional, we can also validate the original address of the sender (the client that created the message)
-        // this could be part of the /directCommand payload for instance
-        // console.log(
-        //  'Got message from ' + senderAddress + ' secrets: ' + decryptedPayload.message
-        // )
-        const secrets = JSON.parse(decryptedPayload.message)
-
-        const cipher = crypto
-          .createCipheriv(
-            FILE_ENCRYPTION_ALGORITHM,
-            Buffer.from(secrets.key, 'hex'),
-            Buffer.from(secrets.iv, 'hex')
-          )
-          .setAutoPadding(true)
-
-        headers['Content-Encoding'] = 'aesgcm'
-
-        return {
-          stream: inputStream.stream.pipe(cipher),
-          status: {
-            httpStatus: inputStream.httpStatus,
-            headers
-          }
-        }
-      } else {
-        // Download request is not using encryption!
-        return {
-          stream: inputStream.stream,
-          status: {
-            httpStatus: inputStream.httpStatus,
-            headers
-          }
-        }
-      }
-    } catch (err) {
-      P2P_CONSOLE_LOGGER.logMessageWithEmoji(
-        'Failure executing downloadURL task: ' + err.message,
-        true,
-        GENERIC_EMOJIS.EMOJI_CROSS_MARK,
-        LOG_LEVELS_STR.LEVEL_ERROR
+  async handle(task: any): Promise<P2PCommandResponse> {
+    if (!this.isDownloadCommand(task)) {
+      throw new Error(
+        `Task has nor DownloadCommand type, nor DownloadUrlCommand type. It has ${typeof task}`
       )
-      return {
-        stream: null,
-        status: { httpStatus: 501, error: 'Unknown error: ' + err.message }
-      }
     }
-  }
-
-  async handle(): Promise<P2PCommandResponse> {
-    const task = this.getTask()
     const node = this.getP2PNode()
     P2P_CONSOLE_LOGGER.logMessage(
       'Download Request recieved with arguments: ' +
@@ -153,15 +139,7 @@ export class DownloadHandler extends Handler {
       true
     )
     // 1. Get the DDO
-    const findDdoTask: any = {
-      command: PROTOCOL_COMMANDS.FIND_DDO,
-      id: task.documentId
-    }
-    const ddo = await new FindDdoHandler(
-      findDdoTask,
-      this.getConfig(),
-      this.getDatabase()
-    ).findAndFormatDdo(task.documentId)
+    const ddo = await new FindDdoHandler(node).findAndFormatDdo(task.documentId)
 
     if (ddo) {
       P2P_CONSOLE_LOGGER.logMessage('DDO for asset found: ' + ddo, true)
@@ -290,13 +268,14 @@ export class DownloadHandler extends Handler {
       const decryptedFilesString = Buffer.from(decryptedUrlBytes).toString()
       const decryptedFileArray = JSON.parse(decryptedFilesString)
       // 7. Proceed to download the file
-      this.setTask({
-        command: PROTOCOL_COMMANDS.DOWNLOAD_URL,
-        fileObject: decryptedFileArray.files[task.fileIndex],
-        aes_encrypted_key: task.aes_encrypted_key
-      })
 
-      return await this.handleDownloadUrlCommand()
+      return await handleDownloadUrlCommand(
+        {
+          fileObject: decryptedFileArray.files[task.fileIndex],
+          aes_encrypted_key: task.aes_encrypted_key
+        },
+        node.getConfig()
+      )
     } catch (e) {
       P2P_CONSOLE_LOGGER.logMessage('decryption error' + e, true)
       return {
